+++ Backend/src/main/java/com/gist/cithub/backend/Tools/entity/ModelsEntity.java
@@ -42,7 +42,6 @@ public class ModelsEntity implements Serializable {
 	 * 
 	 */
 	private Integer userid;
-	private Integer strength;
 
 	private Date createdtime;
