+++ Backend/src/main/resources/mapper/TestSuitesDao.xml
@@ -15,6 +15,7 @@
         <result property="createdtime" column="createdtime"/>
         <result property="lastupdatedtime" column="lastupdatedtime"/>
         <result property="algorithm" column="algorithm"/>
+        <result property="strength" column="strength"/>
     </resultMap>
