+++ Frontend/src/ToolsViews/components/TestSuitesTable.vue
@@ -1,17 +1,55 @@
 <template>
   <div class="card">
     <div class="card-header pb-0">
-      <h5>{{ testSuitesStore.testSuitesList.length }} test suites (strength {{ currentModel.currentModel.strength }})
-        generated by the above model found.</h5>
+      <h5 v-if="props.index">Model #{{ props.index + 1 }}</h5>
+      <h5 v-if="route.query.index">Model #{{ parseInt(route.query.index) + 1 }}</h5>
+      <h6>Model Name: {{ props.model.modelname }}</h6>
+      <h7>Model Description: {{ props.model.modeldescriptions }}</h7>
+      <h7>{{ props.model.NumOfTestSuites }} testsuties found.</h7>
+
+      <!-- <div class="row">
+
+        <div class="col-2">
+          <h6>Model Name</h6>
+          <argon-input v-model="props.model.modelname" type="text" :disabled="'disabled'" />
+        </div>
+
+        <div class="col-2">
+          <h6> Num of params</h6>
+          <argon-input v-model="props.model.NumofParams" type="text" :disabled="'disabled'" />
+        </div>
+
+        <div class="col-2">
+          <h6> Num of Cons</h6>
+          <argon-input v-model="props.model.NumofCons" type="text" :disabled="'disabled'" />
+        </div>
+
+      </div>
+
+      <div class="row">
+        <div class="col-12">
+          <h6>Model Description</h6>
+          <argon-input v-model="props.model.modeldescriptions" type="text" :disabled="'disabled'" />
+        </div>
+      </div>
+
+      <div class="row">
+        <h6> Operation</h6>
+
+        <div class="col-2">
+          <ArgonButton full-width color="primary" variant="gradient" @click="EnterModels(currentModel.currentModel)">
+            <span class="fas fa-book-open text-white me-2" />
+            Edit Model
+          </ArgonButton>
+        </div>
+
+      </div> -->
+
+
     </div>
 
-    <div style="width:100%;margin-bottom: 5px;text-align: center;">
 
-      <ArgonButton style="float: right;margin-right: 3%;" color="primary" variant="gradient" @click="showdialogNew">
-        <span class="ni ni-fat-add ni-lg me-1" />
-        Generate New Test Suites
-      </ArgonButton>
-    </div>
+
 
 
     <!-- New test suites -->
@@ -44,87 +82,96 @@
         <table class="table align-items-center mb-0">
           <thead>
             <tr>
-              <th class="text-uppercase text-secondary text-xxs font-weight-bolder opacity-7">Test Suites</th>
-              <th class="text-uppercase text-secondary text-xxs font-weight-bolder opacity-7 ps-2">
+              <th class="text-secondary  font-weight-bolder opacity-7">Test Suites</th>
+              <th class=" text-center  text-secondary  font-weight-bolder opacity-7 ps-2">
                 <ArgonBadge>Algorithm</ArgonBadge>
               </th>
-              <th class="text-center text-uppercase text-secondary text-xxs font-weight-bolder opacity-7">
-                <ArgonBadge color="info">size</ArgonBadge>
+
+              <th class=" text-center  text-secondary  font-weight-bolder opacity-7 ps-2">
+                <ArgonBadge color="danger">Strength</ArgonBadge>
               </th>
-              <th class="text-center text-uppercase text-secondary text-xxs font-weight-bolder opacity-7">
-                <ArgonBadge color="primary">TIME</ArgonBadge>
+              
+              <th class="text-center  text-secondary  font-weight-bolder opacity-7">
+                <ArgonBadge color="info">Size</ArgonBadge>
               </th>
-              <th class="text-center text-uppercase text-secondary text-xxs font-weight-bolder opacity-7">
+              <th class="text-center  text-secondary  font-weight-bolder opacity-7">
+                <ArgonBadge color="primary">Time</ArgonBadge>
+              </th>
+              <th class="text-center  text-secondary  font-weight-bolder opacity-7">
                 created time
               </th>
-              <th class="text-center text-uppercase text-secondary text-xxs font-weight-bolder opacity-7">
+              <th class="text-center  text-secondary  font-weight-bolder opacity-7">
                 last updated time
               </th>
-              <th class="text-center text-uppercase text-secondary text-xxs font-weight-bolder opacity-7">
+              <th class="text-center  text-secondary  font-weight-bolder opacity-7">
                 Operation
               </th>
 
             </tr>
           </thead>
           <tbody>
-            <tr class="cardTR" v-for="(testSuite, index) in testSuitesStore.testSuitesList">
-
+            <tr class="cardTR" v-for="(testSuite, index) in props.testSuites">
               <td>
-                <div class="d-flex px-2 py-1">
+                <div class="d-flex px-3 py-1">
                   <div class="d-flex flex-column justify-content-center">
-                    <h6 class="mb-0 text-sm">{{ testSuite.testsuitesname }}</h6>
+                    <h6 class=" ">{{ testSuite.testsuitesname }}</h6>
                     <p class="text-xs text-secondary mb-0">{{ testSuite.testsuitesdescriptions }}</p>
                   </div>
                 </div>
               </td>
               <td>
-                <p class="text-xs font-weight-bold mb-0">{{ testSuite.algorithm }}</p>
+                <p class=" text-center font-weight-bold mb-0">{{ testSuite.algorithm }}</p>
+              </td>
+
+              <td>
+                <p class=" text-center font-weight-bold mb-0">{{ testSuite.strength }}</p>
+              </td>
+
+              <td class="align-middle text-center text-sm">
+                <p class=" font-weight-bold mb-0">{{ testSuite.size }}</p>
               </td>
               <td class="align-middle text-center text-sm">
-                <p class="text-xs font-weight-bold mb-0">{{ testSuite.size }}</p>
+                <p class=" font-weight-bold mb-0">{{ testSuite.time }}</p>
               </td>
               <td class="align-middle text-center text-sm">
-                <p class="text-xs font-weight-bold mb-0">{{ testSuite.time }}</p>
-              </td>
-              <td class="align-middle text-center text-sm">
-                <p class="text-xs font-weight-bold mb-0">{{ testSuite.createdtimeFortmat }}</p>
+                <p class=" font-weight-bold mb-0">{{ testSuite.createdtimeFormat }}</p>
               </td>
               <td class="align-middle text-center">
-                <p class="text-xs font-weight-bold mb-0">{{ testSuite.lastupdatedtimeFortmat }}</p>
+                <p class=" font-weight-bold mb-0">{{ testSuite.lastupdatedtimeFormat }}</p>
               </td>
-              <td class="align-middle">
-                <a class="btn btn-link text-dark px-3 mb-0" @click="EnterTestSuiteDetails(testSuite)">
-                  <i class="fas fa-book-open text-primary me-2" aria-hidden="true"></i>Details
-                </a>
+              <td class="align-middle text-center">
 
-                <!-- <a class="btn btn-link text-dark px-3 mb-0" @click="EnterTestSuite(testSuite)">
-                  <i class="fas fa-book-reader text-success me-2" aria-hidden="true"></i>Evaluation
-                </a>
+                <div class="d-flex justify-content-center">
 
-                <a class="btn btn-link text-dark px-3 mb-0" @click="EnterTestSuite(testSuite)">
-                  <i class="fas fa-blender-phone text-info me-2" aria-hidden="true"></i>Prioritisation
-                </a>
+                  <div class="mx-2">
+                    <argon-button @click="EnterTestSuiteDetails(testSuite, index)" color="success" variant="gradient">
+                      <i class="fa fa-thumbtack me-2" aria-hidden="true"></i>
+                      {{ getActionButtonText(route.name) }}
+                    </argon-button>
 
-                <a class="btn btn-link text-dark px-3 mb-0" @click="EnterTestSuite(testSuite)">
-                  <i class="fas fa-caravan text-warning me-2" aria-hidden="true"></i>Reduction
-                </a> -->
+                  </div>
+                  <div>
+                    <el-popconfirm title="Are you sure to delete this project?" confirm-button-text="Yes"
+                      @confirm="confirmDelete(testSuite)">
+                      <template #reference>
+                        <argon-button color="danger" variant="gradient"><i class="far fa-trash-alt me-2"
+                            aria-hidden="true"></i>Delete</argon-button>
+                      </template>
+                    </el-popconfirm>
+                  </div>
+                </div>
 
-
-
-
-                <el-popconfirm title="Are you sure to delete this project?" confirm-button-text="Yes"
-                  @confirm="confirmDelete(testSuite)">
-                  <template #reference>
-                    <a class="btn btn-link text-danger text-gradient px-3 mb-0">
-                      <i class="far fa-trash-alt me-2" aria-hidden="true"></i>Delete
-                    </a>
-                  </template>
-                </el-popconfirm>
               </td>
             </tr>
 
           </tbody>
         </table>
+        <div class="col-12 px-4 my-4">
+          <ArgonButton full-width color="success" variant="gradient" @click="showdialogNew">
+            <span class="ni ni-fat-add ni-lg me-1" />
+            Generate New TestSuite
+          </ArgonButton>
+        </div>
       </div>
     </div>
   </div>
@@ -135,52 +182,28 @@
 import { onMounted, reactive, ref, computed, defineProps } from 'vue';
 import ArgonBadge from '../../CustomizedComponents/ArgonBadge.vue'
 import ArgonButton from '../../CustomizedComponents/ArgonButton.vue';
-import { useModelsStore } from '../../store/modelsStore'
+import ArgonInput from '../../CustomizedComponents/ArgonInput.vue';
+import { useModelsStore } from '../../store/ToolsStore/modelsStore'
 import { routerKey, useRoute, useRouter } from 'vue-router';
 import { request } from '../../request';
 import { ElNotification } from 'element-plus'
-import { useTestSuitesStore } from '../../store/testSuitesStore'
+import { useTestSuitesStore } from '../../store/ToolsStore/testSuitesStore'
 import toolsInfo from "../../CustomizedComponents/tools_info.json";
-import { useCurrentModel } from '../../store/currentModel'
-import { useCurrentTestSuitesStore } from '../../store/currentTestSuite'
+import { useCurrentTestSuitesStore } from '../../store/ToolsStore/currentTestSuite'
+import { useCurrentModel } from '../../store/ToolsStore/currentModel'
 import pinia from '../../store/store'
+import { ElLoading } from 'element-plus'
+
 const router = useRouter();
-const currentModel = useCurrentModel(pinia)
-const testSuitesStore = useTestSuitesStore(pinia)
-const currentTestSuite = useCurrentTestSuitesStore(pinia)
 const route = useRoute()
-
-const listAllTestSuitesByModelID = async () => {
-  try {
-    const testSuitesRes = await request({
-      method: "POST",
-      url: '/tools/testSuites/listTestSuitesByModelID',
-      data: {
-        modelid: route.query.modelid
-      }
-    });
-    testSuitesStore.testSuitesList = testSuitesRes.TestSuites
-
-
-    for (let i = 0; i < testSuitesStore.testSuitesList.length; i++) {
-
-      const timestamp_created = testSuitesStore.testSuitesList[i].createdtime
-      const timestamp_lastupdated = testSuitesStore.testSuitesList[i].lastupdatedtime
-      const dateObject_created = new Date(timestamp_created);
-      const dateObject_lastupdated = new Date(timestamp_lastupdated);
-
-      // 获取可读的时间字符串
-      testSuitesStore.testSuitesList[i].createdtimeFortmat = dateObject_created.toLocaleString();
-      testSuitesStore.testSuitesList[i].lastupdatedtimeFortmat = dateObject_lastupdated.toLocaleString();
-    }
-
-
-
-  } catch (error) {
-    console.error("发生错误", error);
-  }
-
-};
+const props = defineProps({
+  model: Object,
+  testSuites: Object,
+  index: Number
+})
+const testSuitesStore = useTestSuitesStore(pinia)
+const currentModel = useCurrentModel(pinia)
+const currentTestSuite = useCurrentTestSuitesStore(pinia)
 
 const dialogFormVisibleNew = ref(false)
 const dialogformNewTestSuites = reactive({
@@ -192,7 +215,56 @@ const dialogformNewTestSuites = reactive({
 })
 const showdialogNew = () => {
   dialogFormVisibleNew.value = true
-  dialogformNewTestSuites.modelid = route.query.modelid
+}
+
+const getActionButtonText = (routeName) => {
+  switch (routeName) {
+    case 'TestSuites_Home':
+      return 'Enter';
+    case 'TestSuiteDetails':
+      return 'Choose';
+    // 更多情况省略
+    default:
+      return 'Enter';
+  }
+};
+
+const EnterTestSuiteDetails = (testsuite, index) => {
+  // console.log("testsuite",testsuite)
+  let loadingInstance = ElLoading.service({ fullscreen: true })
+
+  currentTestSuite.currentTestSuites.testsuitesid = testsuite.testsuitesid
+  currentTestSuite.currentTestSuites.testsuitesname = testsuite.testsuitesname
+  currentTestSuite.currentTestSuites.testsuitesdescriptions = testsuite.testsuitesdescriptions
+  currentTestSuite.currentTestSuites.testsuitescontents = testsuite.testsuitescontents
+  currentTestSuite.currentTestSuites.algorithm = testsuite.algorithm
+  currentTestSuite.currentTestSuites.strength = testsuite.strength
+  currentTestSuite.currentTestSuites.modelid = testsuite.modelid
+  currentTestSuite.currentTestSuites.size = testsuite.size
+  currentTestSuite.currentTestSuites.time = testsuite.time
+  currentTestSuite.currentTestSuites.createdtime = testsuite.createdtime
+  currentTestSuite.currentTestSuites.lastupdatedtime = testsuite.lastupdatedtime
+  currentTestSuite.currentTestSuites.createdtimeFortmat = testsuite.createdtimeFortmat
+  currentTestSuite.currentTestSuites.lastupdatedtimeFortmat = testsuite.lastupdatedtimeFortmat
+
+  // 在 TestSuites_Home 才 切换页面，在 Details 页面 只切换数据 ，不切换页面
+  if (route.name == 'TestSuites_Home') {
+    router.push(
+      {
+        path: '/tools/TestSuiteDetails',
+        query: {
+          testsuitesid: currentTestSuite.currentTestSuites.testsuitesid,
+          modelid: currentTestSuite.currentTestSuites.modelid,
+          index:index
+        }
+      }
+    )
+  }
+
+  // console.log("route", route.name)
+  loadingInstance.close()
+
+
 }
 
 const confirmGenerateNewTestSuites = async () => {
@@ -201,10 +273,10 @@ const confirmGenerateNewTestSuites = async () => {
   dialogformNewTestSuites.lastupdatedtime = currentDate
   dialogformNewTestSuites.createdtime = currentDate
 
-  if (currentModel.currentModel.modelname == "") {
+  if (props.model.modelname == "") {
     ElNotification({
       title: 'New TestSuites fail!',
-      message:'Model Name can not be empty!',
+      message: 'Model Name can not be empty!',
       type: 'error',
     })
   }
@@ -215,14 +287,14 @@ const confirmGenerateNewTestSuites = async () => {
         try {
           const TestSuitesRes = await request({
             // 这里记得改回去，在校外无法用校内服务器
-            url:tool.url ,
+            url: tool.url,
             // url: 'http://localhost:8300',
             method: 'POST',
             // 注意这里headers一定要加上，不然data末尾会出现莫名其妙的:
             headers: {
               'Content-Type': 'text/plain'
             },
-            data: currentModel.currentModel.modelCithub
+            data: props.model.modelCithub
           })
 
 
@@ -306,169 +378,27 @@ const confirmDelete = (testSuites) => {
     })
   })
 }
-const AlgorithmChosed = ref('')
-const AlgorithmOptions = reactive([])
-const listAllGenerationAlgorithm = () => {
-  for (const tool of toolsInfo.RECORDS) {
-    if (tool.type == "Generation" && !tool.title.includes("GUI")) {
-      AlgorithmOptions.push({ "value": tool.title, "label": tool.title, "url": tool.url })
-    }
-  }
-  // console.log("AlgorithmOptions",AlgorithmOptions)
-}
-const findPosition = (parameterName, value, tableData) => {
-  for (let i = 0; i < tableData.length; i++) {
-    const parameter = tableData[i].Parameter;
-    const values = tableData[i].Value.split(',');
 
-    if (parameter === parameterName) {
-      const valueIndex = values.indexOf(value);
-      if (valueIndex !== -1) {
-        return { parameterIndex: i, valueIndex };
-      }
-    }
-  }
-
-  return { parameterIndex: -1, valueIndex: -1 };
-}
-const loadModel = (model) => {
-  // 解析参数和参数取值的Json字符串
-  const parsedData = JSON.parse(model.paramsvalues)
-  // 移除 row_index 属性
-  const tableDataTmp = parsedData.map(item => {
-    const { row_index, ...rest } = item;
-    return rest;
-  });
-
-  // 加载约束table
-  let tempArray = []
-  let param_count = 0
-  for (let i = 0, len = tableDataTmp.length; i < len; i++) {
-    if (tableDataTmp[i].Value != '') { tempArray.push(tableDataTmp[i].Value.split(',').length) }
-    if (tableDataTmp[i].Parameter != '') { param_count = param_count + 1 }
-
-  }
-  let modelObject = {
-    system: "",
-    strength: '',
-    parameter: '',
-    values: '',
-    constraints: []
-
-  }
-  // 统计模型基本数据
-  if (model.modelname) { modelObject.system = model.modelname }
-
-  modelObject.strength = model.strength
-  modelObject.parameter = param_count
-  modelObject.values = JSON.stringify(tempArray)
-
-
-  // 对 tableData 的数据做处理，转成被constraintsTableData
-  let tempArrayCons = []
-  for (let i = 0, len = tableDataTmp.length; i < len; i++) {
-    let tempObj = {}
-    tempObj.Parameter = tableDataTmp[i].Parameter
-    tempObj.valueArray = tableDataTmp[i].Value.split(',').map((value, index) => ({
-      [`Value${index + 1}`]: value
-    }));
-    tempObj.ValueDomain = tempObj.valueArray.length
-    tempArrayCons.push(tempObj)
-
-  }
-
-
-  let consArray = JSON.parse(model.cons)
-  // 对每个约束进行处理，将其转换成 '参数索引'/'取值索引' 的形式
-  let consArrayToAPI = []
-  for (const constraint of consArray) {
-    let consArrayTemp = []
-    // 遍历元素中的键值对
-    for (const key in constraint) {
-      if (Object.hasOwnProperty.call(constraint, key)) {
-        const elements = constraint[key];
-        // elements 是一个数组，包含了多个键值对
-        for (const element of elements) {
-          // 在这里访问 Parameter 和 Value
-          const parameter = element.Parameter;
-          const value = element.Value;
-
-          // 这里可以使用 parameter 和 value 进行其他操作
-          // console.log(`Key: ${key}, Parameter: ${parameter}, Value: ${value}`);
-          let constempObj = findPosition(parameter, value, tableDataTmp)
-          let consString = `\'${constempObj.parameterIndex}/${constempObj.valueIndex}\'`
-          consArrayTemp.push(consString)
-        }
-      }
-      consArrayToAPI.push(consArrayTemp)
-    }
-  }
-  modelObject.constraints = consArrayToAPI
-
-  // 将模型数据转换成cithub格式
-  currentModel.currentModel.modelCithub = JSON.stringify(modelObject, null, 6).replace(/"/g, '')
-}
-
-const model = reactive({})
-const listModelInfoByModelID = async () => {
-  try {
-    const res = await request({
-      url: '/tools/models/info/' + route.query.modelid,
-      method: 'POST'
-    });
-
-    model.modelid = res.models.modelid
-    model.strength = res.models.strength
-    model.modelname = res.models.modelname
-    model.modeldescriptions = res.models.modeldescriptions
-    model.paramsvalues = res.models.paramsvalues
-    model.cons = res.models.cons
-    model.lastupdatedtime = res.models.lastupdatedtime
-    model.createdtime = res.models.createdtime
-
-    currentModel.currentModel = {}
-    currentModel.currentModel = model
-    loadModel(model);
-  } catch (err) {
-    console.error(err);
-  }
-}
-
-const EnterTestSuiteDetails = (testsuite) => {
-  currentTestSuite.currentTestSuites.testsuitesid = testsuite.testsuitesid
-  currentTestSuite.currentTestSuites.testsuitesname = testsuite.testsuitesname
-  currentTestSuite.currentTestSuites.testsuitesdescriptions = testsuite.testsuitesdescriptions
-  currentTestSuite.currentTestSuites.testsuitescontents = testsuite.testsuitescontents
-  currentTestSuite.currentTestSuites.algorithm = testsuite.algorithm
-  currentTestSuite.currentTestSuites.modelid = testsuite.modelid
-  currentTestSuite.currentTestSuites.size = testsuite.size
-  currentTestSuite.currentTestSuites.time = testsuite.time
-  currentTestSuite.currentTestSuites.createdtime = testsuite.createdtime
-  currentTestSuite.currentTestSuites.lastupdatedtime = testsuite.lastupdatedtime
-  currentTestSuite.currentTestSuites.createdtimeFortmat = testsuite.createdtimeFortmat
-  currentTestSuite.currentTestSuites.lastupdatedtimeFortmat = testsuite.lastupdatedtimeFortmat
-  router.push(
-    {
-      name: 'TestSuiteDetails',
-      query: {
-        TestSuiteID: currentTestSuite.currentTestSuites.testsuitesid
-      }
-    }
-  )
-
-}
 
 onMounted(async () => {
 
-  // 加载当前模型信息
-  await listModelInfoByModelID()
+  // 让选中的行高亮
+  try {
+    // console.log("TestSuiteTable的数据加载 testsuites", props.testSuites)
+    // console.log("props.model", props.model)
 
-  // 列出该model下所有的ca
-  await listAllTestSuitesByModelID()
+  } catch (error) {
+    console.log("error in TestSuitesTable", error)
 
-  // 加载所有生成算法
-  await listAllGenerationAlgorithm()
-
+  }
 
 })
 </script>
+
+
+<style scoped>
+.highlighted-row {
+  border: 2px solid red;
+  /* 定义高亮样式 */
+}
+</style>
