+++ Frontend/src/ToolsViews/components/_PrioritisationTable.vue
