+++ Frontend/src/ToolsViews/components/_ProjectCard.vue
