+++ Frontend/src/ToolsViews/components/_ReductionTable.vue
