+++ Frontend/src/ToolsViews/components/_TestSuitesTable.vue
@@ -1,0 +1,474 @@
+<template>
+  <div class="card">
+    <div class="card-header pb-0">
+      <h5>{{ testSuitesStore.testSuitesList.length }} test suites (strength {{ currentModel.currentModel.strength }})
+        generated by the above model found.</h5>
+    </div>
+
+    <div style="width:100%;margin-bottom: 5px;text-align: center;">
+
+      <ArgonButton style="float: right;margin-right: 3%;" color="primary" variant="gradient" @click="showdialogNew">
+        <span class="ni ni-fat-add ni-lg me-1" />
+        Generate New Test Suites
+      </ArgonButton>
+    </div>
+
+
+    <!-- New test suites -->
+    <el-dialog v-model="dialogFormVisibleNew" title="New Test Suite">
+      <el-form :model="dialogformNewTestSuites" label-position="right" label-width="170px">
+
+        <el-form-item label="Test Suite Name:">
+          <el-input v-model="dialogformNewTestSuites.testsuitesname" />
+        </el-form-item>
+        <el-form-item label="Test Suite Description:">
+          <el-input autosize type="textarea" v-model="dialogformNewTestSuites.testsuitesdescriptions" />
+        </el-form-item>
+        <el-form-item label="Algorithm: ">
+          <el-select v-model="AlgorithmChosed" class="m-2" placeholder="Select an Algorithm for generating">
+            <el-option v-for="item in AlgorithmOptions" :key="item.value" :label="item.label" :value="item.value" />
+          </el-select>
+        </el-form-item>
+      </el-form>
+      <template #footer>
+        <span class="dialog-footer">
+          <el-button @click="dialogFormVisibleNew = false">Cancel</el-button>
+          <el-button type="primary" @click="confirmGenerateNewTestSuites">
+            Confirm
+          </el-button>
+        </span>
+      </template>
+    </el-dialog>
+    <div class="card-body px-0 pt-0 pb-2">
+      <div class="table-responsive p-0">
+        <table class="table align-items-center mb-0">
+          <thead>
+            <tr>
+              <th class="text-uppercase text-secondary text-xxs font-weight-bolder opacity-7">Test Suites</th>
+              <th class="text-uppercase text-secondary text-xxs font-weight-bolder opacity-7 ps-2">
+                <ArgonBadge>Algorithm</ArgonBadge>
+              </th>
+              <th class="text-center text-uppercase text-secondary text-xxs font-weight-bolder opacity-7">
+                <ArgonBadge color="info">size</ArgonBadge>
+              </th>
+              <th class="text-center text-uppercase text-secondary text-xxs font-weight-bolder opacity-7">
+                <ArgonBadge color="primary">TIME</ArgonBadge>
+              </th>
+              <th class="text-center text-uppercase text-secondary text-xxs font-weight-bolder opacity-7">
+                created time
+              </th>
+              <th class="text-center text-uppercase text-secondary text-xxs font-weight-bolder opacity-7">
+                last updated time
+              </th>
+              <th class="text-center text-uppercase text-secondary text-xxs font-weight-bolder opacity-7">
+                Operation
+              </th>
+
+            </tr>
+          </thead>
+          <tbody>
+            <tr class="cardTR" v-for="(testSuite, index) in testSuitesStore.testSuitesList">
+
+              <td>
+                <div class="d-flex px-2 py-1">
+                  <div class="d-flex flex-column justify-content-center">
+                    <h6 class="mb-0 text-sm">{{ testSuite.testsuitesname }}</h6>
+                    <p class="text-xs text-secondary mb-0">{{ testSuite.testsuitesdescriptions }}</p>
+                  </div>
+                </div>
+              </td>
+              <td>
+                <p class="text-xs font-weight-bold mb-0">{{ testSuite.algorithm }}</p>
+              </td>
+              <td class="align-middle text-center text-sm">
+                <p class="text-xs font-weight-bold mb-0">{{ testSuite.size }}</p>
+              </td>
+              <td class="align-middle text-center text-sm">
+                <p class="text-xs font-weight-bold mb-0">{{ testSuite.time }}</p>
+              </td>
+              <td class="align-middle text-center text-sm">
+                <p class="text-xs font-weight-bold mb-0">{{ testSuite.createdtimeFortmat }}</p>
+              </td>
+              <td class="align-middle text-center">
+                <p class="text-xs font-weight-bold mb-0">{{ testSuite.lastupdatedtimeFortmat }}</p>
+              </td>
+              <td class="align-middle">
+                <a class="btn btn-link text-dark px-3 mb-0" @click="EnterTestSuiteDetails(testSuite)">
+                  <i class="fas fa-book-open text-primary me-2" aria-hidden="true"></i>Details
+                </a>
+
+                <!-- <a class="btn btn-link text-dark px-3 mb-0" @click="EnterTestSuite(testSuite)">
+                  <i class="fas fa-book-reader text-success me-2" aria-hidden="true"></i>Evaluation
+                </a>
+
+                <a class="btn btn-link text-dark px-3 mb-0" @click="EnterTestSuite(testSuite)">
+                  <i class="fas fa-blender-phone text-info me-2" aria-hidden="true"></i>Prioritisation
+                </a>
+
+                <a class="btn btn-link text-dark px-3 mb-0" @click="EnterTestSuite(testSuite)">
+                  <i class="fas fa-caravan text-warning me-2" aria-hidden="true"></i>Reduction
+                </a> -->
+
+
+
+
+                <el-popconfirm title="Are you sure to delete this project?" confirm-button-text="Yes"
+                  @confirm="confirmDelete(testSuite)">
+                  <template #reference>
+                    <a class="btn btn-link text-danger text-gradient px-3 mb-0">
+                      <i class="far fa-trash-alt me-2" aria-hidden="true"></i>Delete
+                    </a>
+                  </template>
+                </el-popconfirm>
+              </td>
+            </tr>
+
+          </tbody>
+        </table>
+      </div>
+    </div>
+  </div>
+</template>
+
+<script setup>
+
+import { onMounted, reactive, ref, computed, defineProps } from 'vue';
+import ArgonBadge from '../../CustomizedComponents/ArgonBadge.vue'
+import ArgonButton from '../../CustomizedComponents/ArgonButton.vue';
+import { useModelsStore } from '../../store/modelsStore'
+import { routerKey, useRoute, useRouter } from 'vue-router';
+import { request } from '../../request';
+import { ElNotification } from 'element-plus'
+import { useTestSuitesStore } from '../../store/testSuitesStore'
+import toolsInfo from "../../CustomizedComponents/tools_info.json";
+import { useCurrentModel } from '../../store/currentModel'
+import { useCurrentTestSuitesStore } from '../../store/currentTestSuite'
+import pinia from '../../store/store'
+const router = useRouter();
+const currentModel = useCurrentModel(pinia)
+const testSuitesStore = useTestSuitesStore(pinia)
+const currentTestSuite = useCurrentTestSuitesStore(pinia)
+const route = useRoute()
+
+const listAllTestSuitesByModelID = async () => {
+  try {
+    const testSuitesRes = await request({
+      method: "POST",
+      url: '/tools/testSuites/listTestSuitesByModelID',
+      data: {
+        modelid: route.query.modelid
+      }
+    });
+    testSuitesStore.testSuitesList = testSuitesRes.TestSuites
+
+
+    for (let i = 0; i < testSuitesStore.testSuitesList.length; i++) {
+
+      const timestamp_created = testSuitesStore.testSuitesList[i].createdtime
+      const timestamp_lastupdated = testSuitesStore.testSuitesList[i].lastupdatedtime
+      const dateObject_created = new Date(timestamp_created);
+      const dateObject_lastupdated = new Date(timestamp_lastupdated);
+
+      // 获取可读的时间字符串
+      testSuitesStore.testSuitesList[i].createdtimeFortmat = dateObject_created.toLocaleString();
+      testSuitesStore.testSuitesList[i].lastupdatedtimeFortmat = dateObject_lastupdated.toLocaleString();
+    }
+
+
+
+  } catch (error) {
+    console.error("发生错误", error);
+  }
+
+};
+
+const dialogFormVisibleNew = ref(false)
+const dialogformNewTestSuites = reactive({
+  testsuitesname: '',
+  testsuitesdescriptions: '',
+  lastupdatedtime: '',
+  createdtime: '',
+  modelid: route.query.modelid
+})
+const showdialogNew = () => {
+  dialogFormVisibleNew.value = true
+  dialogformNewTestSuites.modelid = route.query.modelid
+}
+
+const confirmGenerateNewTestSuites = async () => {
+  // 获取当前时刻的Date对象
+  const currentDate = new Date();
+  dialogformNewTestSuites.lastupdatedtime = currentDate
+  dialogformNewTestSuites.createdtime = currentDate
+
+  if (currentModel.currentModel.modelname == "") {
+    ElNotification({
+      title: 'New TestSuites fail!',
+      message:'Model Name can not be empty!',
+      type: 'error',
+    })
+  }
+  else {
+    // 发送请求，对 model 得到 test suites
+    for (const tool of AlgorithmOptions) {
+      if (tool.value == AlgorithmChosed.value) {
+        try {
+          const TestSuitesRes = await request({
+            // 这里记得改回去，在校外无法用校内服务器
+            url:tool.url ,
+            // url: 'http://localhost:8300',
+            method: 'POST',
+            // 注意这里headers一定要加上，不然data末尾会出现莫名其妙的:
+            headers: {
+              'Content-Type': 'text/plain'
+            },
+            data: currentModel.currentModel.modelCithub
+          })
+
+
+          const newTestSuitesRes = await request({
+            url: "/tools/testSuites/NewTestSuites",
+            method: "POST",
+            data:
+            {
+              testsuitesname: dialogformNewTestSuites.testsuitesname,
+              testsuitesdescriptions: dialogformNewTestSuites.testsuitesdescriptions,
+              lastupdatedtime: dialogformNewTestSuites.lastupdatedtime,
+              createdtime: dialogformNewTestSuites.createdtime,
+              modelid: dialogformNewTestSuites.modelid,
+              testsuitescontents: JSON.stringify(TestSuitesRes),
+              time: TestSuitesRes.time,
+              size: TestSuitesRes.size,
+              algorithm: AlgorithmChosed.value
+            }
+          })
+          // console.log("newTestSuitesRes", newTestSuitesRes)
+
+          if (newTestSuitesRes.NewStatus == 'success!') {
+
+            listAllTestSuitesByModelID()
+            ElNotification({
+              title: 'Save Success!',
+              type: 'success',
+            })
+            dialogFormVisibleNew.value = false
+          }
+          else {
+            ElNotification({
+              title: 'Save fail!',
+              type: 'error',
+            })
+            dialogFormVisibleNew.value = false
+
+          }
+        }
+        catch (err) {
+          console.log("err", err)
+          ElNotification({
+            title: 'Save fail!',
+            type: 'error',
+          })
+          dialogFormVisibleNew.value = false
+
+
+        }
+      }
+    }
+  }
+
+
+}
+const confirmDelete = (testSuites) => {
+  // console.log("testSuites", testSuites)
+  request({
+    url: '/tools/testSuites/DeleteByTestSuitesID',
+    method: 'POST',
+    data: {
+      testsuitesid: testSuites.testsuitesid
+    }
+  }).then((res) => {
+    if (res.DeleteStatus == 'success!') {
+      listAllTestSuitesByModelID()
+      // 实时更新页面数据
+      ElNotification({
+        title: 'Delete Success!',
+        message: 'please check the results',
+        type: 'success',
+      })
+
+    }
+  }).catch((error) => {
+    // console.log(error)
+    ElNotification({
+      title: 'Delete Error!',
+      message: 'please check the results',
+      type: 'error',
+    })
+  })
+}
+const AlgorithmChosed = ref('')
+const AlgorithmOptions = reactive([])
+const listAllGenerationAlgorithm = () => {
+  for (const tool of toolsInfo.RECORDS) {
+    if (tool.type == "Generation" && !tool.title.includes("GUI")) {
+      AlgorithmOptions.push({ "value": tool.title, "label": tool.title, "url": tool.url })
+    }
+  }
+  // console.log("AlgorithmOptions",AlgorithmOptions)
+}
+const findPosition = (parameterName, value, tableData) => {
+  for (let i = 0; i < tableData.length; i++) {
+    const parameter = tableData[i].Parameter;
+    const values = tableData[i].Value.split(',');
+
+    if (parameter === parameterName) {
+      const valueIndex = values.indexOf(value);
+      if (valueIndex !== -1) {
+        return { parameterIndex: i, valueIndex };
+      }
+    }
+  }
+
+  return { parameterIndex: -1, valueIndex: -1 };
+}
+const loadModel = (model) => {
+  // 解析参数和参数取值的Json字符串
+  const parsedData = JSON.parse(model.paramsvalues)
+  // 移除 row_index 属性
+  const tableDataTmp = parsedData.map(item => {
+    const { row_index, ...rest } = item;
+    return rest;
+  });
+
+  // 加载约束table
+  let tempArray = []
+  let param_count = 0
+  for (let i = 0, len = tableDataTmp.length; i < len; i++) {
+    if (tableDataTmp[i].Value != '') { tempArray.push(tableDataTmp[i].Value.split(',').length) }
+    if (tableDataTmp[i].Parameter != '') { param_count = param_count + 1 }
+
+  }
+  let modelObject = {
+    system: "",
+    strength: '',
+    parameter: '',
+    values: '',
+    constraints: []
+
+  }
+  // 统计模型基本数据
+  if (model.modelname) { modelObject.system = model.modelname }
+
+  modelObject.strength = model.strength
+  modelObject.parameter = param_count
+  modelObject.values = JSON.stringify(tempArray)
+
+
+  // 对 tableData 的数据做处理，转成被constraintsTableData
+  let tempArrayCons = []
+  for (let i = 0, len = tableDataTmp.length; i < len; i++) {
+    let tempObj = {}
+    tempObj.Parameter = tableDataTmp[i].Parameter
+    tempObj.valueArray = tableDataTmp[i].Value.split(',').map((value, index) => ({
+      [`Value${index + 1}`]: value
+    }));
+    tempObj.ValueDomain = tempObj.valueArray.length
+    tempArrayCons.push(tempObj)
+
+  }
+
+
+  let consArray = JSON.parse(model.cons)
+  // 对每个约束进行处理，将其转换成 '参数索引'/'取值索引' 的形式
+  let consArrayToAPI = []
+  for (const constraint of consArray) {
+    let consArrayTemp = []
+    // 遍历元素中的键值对
+    for (const key in constraint) {
+      if (Object.hasOwnProperty.call(constraint, key)) {
+        const elements = constraint[key];
+        // elements 是一个数组，包含了多个键值对
+        for (const element of elements) {
+          // 在这里访问 Parameter 和 Value
+          const parameter = element.Parameter;
+          const value = element.Value;
+
+          // 这里可以使用 parameter 和 value 进行其他操作
+          // console.log(`Key: ${key}, Parameter: ${parameter}, Value: ${value}`);
+          let constempObj = findPosition(parameter, value, tableDataTmp)
+          let consString = `\'${constempObj.parameterIndex}/${constempObj.valueIndex}\'`
+          consArrayTemp.push(consString)
+        }
+      }
+      consArrayToAPI.push(consArrayTemp)
+    }
+  }
+  modelObject.constraints = consArrayToAPI
+
+  // 将模型数据转换成cithub格式
+  currentModel.currentModel.modelCithub = JSON.stringify(modelObject, null, 6).replace(/"/g, '')
+}
+
+const model = reactive({})
+const listModelInfoByModelID = async () => {
+  try {
+    const res = await request({
+      url: '/tools/models/info/' + route.query.modelid,
+      method: 'POST'
+    });
+
+    model.modelid = res.models.modelid
+    model.strength = res.models.strength
+    model.modelname = res.models.modelname
+    model.modeldescriptions = res.models.modeldescriptions
+    model.paramsvalues = res.models.paramsvalues
+    model.cons = res.models.cons
+    model.lastupdatedtime = res.models.lastupdatedtime
+    model.createdtime = res.models.createdtime
+
+    currentModel.currentModel = {}
+    currentModel.currentModel = model
+    loadModel(model);
+  } catch (err) {
+    console.error(err);
+  }
+}
+
+const EnterTestSuiteDetails = (testsuite) => {
+  currentTestSuite.currentTestSuites.testsuitesid = testsuite.testsuitesid
+  currentTestSuite.currentTestSuites.testsuitesname = testsuite.testsuitesname
+  currentTestSuite.currentTestSuites.testsuitesdescriptions = testsuite.testsuitesdescriptions
+  currentTestSuite.currentTestSuites.testsuitescontents = testsuite.testsuitescontents
+  currentTestSuite.currentTestSuites.algorithm = testsuite.algorithm
+  currentTestSuite.currentTestSuites.modelid = testsuite.modelid
+  currentTestSuite.currentTestSuites.size = testsuite.size
+  currentTestSuite.currentTestSuites.time = testsuite.time
+  currentTestSuite.currentTestSuites.createdtime = testsuite.createdtime
+  currentTestSuite.currentTestSuites.lastupdatedtime = testsuite.lastupdatedtime
+  currentTestSuite.currentTestSuites.createdtimeFortmat = testsuite.createdtimeFortmat
+  currentTestSuite.currentTestSuites.lastupdatedtimeFortmat = testsuite.lastupdatedtimeFortmat
+  router.push(
+    {
+      name: 'TestSuiteDetails',
+      query: {
+        TestSuiteID: currentTestSuite.currentTestSuites.testsuitesid
+      }
+    }
+  )
+
+}
+
+onMounted(async () => {
+
+  // 加载当前模型信息
+  await listModelInfoByModelID()
+
+  // 列出该model下所有的ca
+  await listAllTestSuitesByModelID()
+
+  // 加载所有生成算法
+  await listAllGenerationAlgorithm()
+
+
+})
+</script>
