+++ Frontend/src/store/ToolsStore/currentModel.js
@@ -2,7 +2,8 @@ import { defineStore } from "pinia";
 export const useCurrentModel = defineStore("CurrentModel", {
     state: () => {
         return {
-            currentModel: {}
+            currentModel: {},
+            currentModelTestSuites:{}
             
         }
     },
