--- Frontend/src/store/conversionStore.js
@@ -1,16 +1,0 @@
-import { defineStore } from "pinia";
-export const useConversionStore = defineStore("ConversionStore", {
-    state: () => {
-        return {
-            ConversionList:[], 
-
-        }
-    },
-    actions: {
-
-    },
-    getters: {
-
-
-    }
-})
