--- Frontend/src/store/currentProject.js
@@ -1,21 +1,0 @@
-import { defineStore } from "pinia";
-export const useCurrentProject = defineStore("CurrentProject", {
-    state: () => {
-        return {
-            projectid:'',
-            projectname:'',
-            userid:'',
-            projectdescriptions:'',
-            createdtime:'',
-            lastupdatedtime:'',
-
-        }
-    },
-    actions: {
-
-    },
-    getters: {
-
-
-    }
-})
