--- Frontend/src/store/prioritisationStore.js
@@ -1,16 +1,0 @@
-import { defineStore } from "pinia";
-export const usePrioritisationStore = defineStore("PrioritisationStore", {
-    state: () => {
-        return {
-            PrioritisationList:[], 
-
-        }
-    },
-    actions: {
-
-    },
-    getters: {
-
-
-    }
-})
