--- Frontend/src/store/projectsStore.js
@@ -1,16 +1,0 @@
-import { defineStore } from "pinia";
-export const useProjectsStore = defineStore("ProjectsStore", {
-    state: () => {
-        return {
-            projectList:[] //该用户所有的project的list
-
-        }
-    },
-    actions: {
-
-    },
-    getters: {
-
-
-    }
-})
