--- Frontend/src/store/reductionStore.js
@@ -1,16 +1,0 @@
-import { defineStore } from "pinia";
-export const useReductionStore = defineStore("ReductionStore", {
-    state: () => {
-        return {
-            ReductionList:[], 
-
-        }
-    },
-    actions: {
-
-    },
-    getters: {
-
-
-    }
-})
