--- Frontend/src/store/testSuitesStore.js
@@ -1,16 +1,0 @@
-import { defineStore } from "pinia";
-export const useTestSuitesStore = defineStore("TestSuitesStore", {
-    state: () => {
-        return {
-            testSuitesList:[], 
-
-        }
-    },
-    actions: {
-
-    },
-    getters: {
-
-
-    }
-})
